@@ -44,10 +44,10 @@
   "module": "angular-star-rating.js",
   "typings": "angular-star-rating.d.ts",
   "peerDependencies": {
-    "@angular/core": "^4.0.0 || ^5.0.0",
-    "@angular/forms": "^4.0.0 || ^5.0.0",
-    "rxjs": "^5.1.0",
-    "zone.js": "^0.8.4",
+    "@angular/core": "^10.0.0 || ^12.0.0",
+    "@angular/forms": "^10.0.0 || ^12.0.0",
+    "rxjs": "^6.6.0",
+    "zone.js": "^0.11.4",
     "css-star-rating": "^1.2.4"
   },
   "ngPackage": {
